CustomSideDrawer applied to searchScreen
